// src/types/Post.ts
// src/types/Post.ts
export type Post = {
  id: string;
  title: string;
  content: string;
  author: string;
  createdAt: any;
  likes: number;
  batonCount: number;
  genre: string;
  thumbnail: string;
  members: number;
  lastAuthor: string;
  postedAgo: string;
  latestEpisode: string;
  excerpt: string;
  comments: number;
};